requirements.txt
selenium
pytest
pytest-html
allure-pytest
appium-python-client
faker
